finished /wa reload code added permissions to /wa update command - currently using wa.reload node found and fixed a small memory leak leaving 3 mysql connections open / in use
git-svn-id: http://webauctionplus.googlecode.com/svn/trunk@236 c41ed8a6-1c15-8a3d-a996-777bd911872f

+++ java/src/me/lorenzop/webauctionplus/tasks/ShoutSignTask.java
@@ -65,7 +65,7 @@ public class ShoutSignTask implements Runnable {
 				}
 				Double xValue = key.getX();
 				Double zValue = key.getZ();
-				int radius = (Integer)plugin.shoutSigns.get(key);
+				int radius = plugin.shoutSigns.get(key);
 				for(Player player : playerList) {
 					Double playerX = player.getLocation().getX();
 					Double playerZ = player.getLocation().getZ();
